pull out debug_print so squelch can build without rtl_airband.[cpp|h]
